fix: ノイズゲートによるEMA崩壊バグを修正 & feat: 上級者向け音声処理パラメータをConfig化
- fix: ノイズゲートで弾いた0値がEMAに入り急速にゼロ化するバグを修正
- fix: 長時間無音後の再発話で古いEMAとブレンドされる問題を修正 (陳腐化リセット)
- fix: 例外時のemaLastUpdateMapクリーンアップ漏れを修正
- feat: AudioUtilsの7定数をConfig上級者設定として外部化
  - lowpass_alpha, highpass_alpha, trim_denominator
  - zcr_noise_threshold, zcr_noise_penalty
  - silence_threshold, scaled_db_max
- refactor: VoiceMeterManagerのMAX_DBをConfig.scaledDbMaxに連動

+++ src/main/java/io/wax100/simplevoicechatinteraction/ShockwaveExecutor.java
@@ -108,7 +108,7 @@ public class ShockwaveExecutor {
         if (overdriveProgress > 0.0) {
             double overdriveMult = 1.0 + (Config.shockwaveOverdriveMultiplier - 1.0) * overdriveProgress;
             radius *= overdriveMult;
-            damage *= overdriveMult;
+            damage *= (float) overdriveMult;
         }
         int darknessDuration = Config.shockwaveDarknessDuration;
